Refactor the Corrector classes into a `lightkurve.correctors` sub-package (#519)
* Refactor the Corrector classes into a sub-package

* fix syntax error
