+++ setup.py
@@ -38,8 +38,9 @@ setup(name='lightkurve',
       license='MIT',
       package_dir={
             'lightkurve': 'lightkurve',
+            'lightkurve.correctors': 'lightkurve/correctors',
             'lightkurve.prf': 'lightkurve/prf'},
-      packages=['lightkurve', 'lightkurve.prf'],
+      packages=['lightkurve', 'lightkurve.correctors', 'lightkurve.prf'],
       install_requires=install_requires,
       extras_require=extras_require,
       setup_requires=['pytest-runner'],
